+++ app/Models/Product.php
@@ -56,7 +56,7 @@ class Product extends Model
     {
         static::deleted(function (Product $product) {
             $image = $product->image;
-            if (Storage::disk("products")->exists($image)) {
+            if ($image && Storage::disk("products")->exists($image)) {
                 Storage::disk("products")->delete($image);
             }
         });
